improve cftal::test::make_f16/float

+++ test/f32_f64.cc
@@ -35,8 +35,8 @@ double cftal::test::make_double(unsigned sgn, unsigned exp, uint64_t sig)
 
 float cftal::test::make_float(unsigned sgn, unsigned exp, uint32_t sig)
 {
-    uint32_t _sgn= uint64_t(sgn & 1) << 31;
-    uint32_t _exp= uint64_t(exp & 0xFF) << 23;
+    uint32_t _sgn= uint32_t(sgn & 1) << 31;
+    uint32_t _exp= uint32_t(exp & 0xFF) << 23;
     uint32_t _sig= sig & 0x007fffff;
     union {
         uint32_t _u;
@@ -48,8 +48,8 @@ float cftal::test::make_float(unsigned sgn, unsigned exp, uint32_t sig)
 
 cftal::f16_t cftal::test::make_f16(unsigned sgn, unsigned exp, uint16_t sig)
 {
-    uint16_t _sgn= uint64_t(sgn & 1) << 15;
-    uint16_t _exp= uint64_t(exp & 0x3F) << 10;
+    uint16_t _sgn= uint16_t(sgn & 1) << 15;
+    uint16_t _exp= uint16_t(exp & 0x3F) << 10;
     uint16_t _sig= sig & 0x003ff;
     union {
         uint16_t _u;
